# defs/api.yaml
openapi: 3.0.0

info:
  title: Sonya WQ
  description: sonya is a fast, distributed queue that provide a flexible realization of the Web Queue in Web Queue Worker architecture.
  contact:
    name: Issues
    url: https://github.com/Mnwa/sonya/issues
  license:
    name: MIT
    url: https://github.com/Mnwa/sonya/blob/master/LICENSE
  version: "0.1"

paths:
  /queue/create/{queue_name}:
    post:
      summary: Create queue.
      description: This method will create queue on all sonya instances
      parameters:
        - in: path
          name: queue_name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'

  /queue/close/{queue_name}:
    post:
      summary: Close queue.
      description: This method will close queue on all sonya instances
      parameters:
        - in: path
          name: queue_name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'

  /queue/send/{queue_name}:
    post:
      summary: Send message to queue.
      description: This method will sent queue on all sonya instances
      parameters:
        - in: path
          name: queue_name
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventMessage'
      responses:
        '200':
          description: success result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /queue/listen/longpoll/{queue_name}/{id}:
    get:
      summary: Longpoll listen queue id.
      description: This method will respond with the first message sent in this queue with this id
      parameters:
        - in: path
          name: queue_name
          schema:
            type: string
          required: true
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMessage'
  /queue/listen/longpoll/{queue_name}:
    get:
      summary: Longpoll listen queue.
      description: This method will respond with the first message sent in this queue
      parameters:
        - in: path
          name: queue_name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMessage'
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    EventMessage:
      type: object
      properties:
        id:
          type: string
          description: your message identificator
        timestamp:
          type: integer
          description: event timestamp
        payload:
          type: object
          description: any object
          additionalProperties: true
      required:
        - id
        - payload